export is normed
