+++ src/export.c
@@ -6,45 +6,11 @@
 /*   By: linhnguy <linhnguy@hive.student.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/02 16:33:21 by linhnguy          #+#    #+#             */
-/*   Updated: 2024/06/04 14:01:56 by linhnguy         ###   ########.fr       */
+/*   Updated: 2024/06/05 17:35:12 by linhnguy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/minishell.h"
-
-void static	swap_strings(char **a, char **b)
-{
-	char	*temp;
-
-	temp = *a;
-	*a = *b;
-	*b = temp;
-}
-
-void	sort_strings(char **arr)
-{
-	int	swapped;
-	int	i;
-	int	size;
-
-	size = array_len(arr);
-	swapped = 1;
-	while (swapped)
-	{
-		swapped = 0;
-		i = 0;
-		while (i < size - 1)
-		{
-			if (ft_strncmp(arr[i], arr[i + 1],
-					ft_strlen(arr[i]) + ft_strlen(arr[i + 1])) > 0)
-			{
-				swap_strings(&arr[i], &arr[i + 1]);
-				swapped = 1;
-			}
-			i++;
-		}
-	}
-}
 
 _Bool	check_key_exist(char *env, char *cmd)
 {
@@ -54,36 +20,6 @@ _Bool	check_key_exist(char *env, char *cmd)
 		&& cmd[ft_strlen(cmd) - 1] == '=')
 		return (0);
 	return (0);
-}
-
-_Bool	update_key(t_pipex *data, char **env, char *cmd)
-{
-	char	*equal;
-	char	**tmp;
-	int		i;
-	int		flag;
-
-	i = 0;
-	flag = 0;
-	tmp = env;
-	while (tmp[i])
-	{
-		if (check_key_exist(tmp[i], cmd))
-			return (1);
-		equal = ft_strchr(tmp[i], '=');
-		if (ft_strncmp(tmp[i], cmd, equal - tmp[i] + 1) == 0
-			|| (!equal && ft_strncmp(tmp[i], cmd, ft_strlen(tmp[i]))
-				== 0 && cmd[ft_strlen(tmp[i])] == '='))
-		{
-			flag = 1;
-			free(tmp[i]);
-			env[i] = ft_strdup(cmd);
-			if (!env[i])
-				set_error_and_print(data, -1, "strdup failed in update_key");
-		}
-		i++;
-	}
-	return (flag);
 }
 
 char	**putstr_in_array(t_pipex *data, char **env, char *str)
@@ -126,41 +62,13 @@ _Bool	check_key(char *str)
 	return (1);
 }
 
-void	print_export(char *env, int fd_out)
-{
-	char		*equal;
-	int			i;
-
-	i = 0;
-	equal = ft_strchr(env, '=');
-	if (!equal)
-		ft_printf(fd_out, "declare -x %s\n", env);
-	else if (*(equal + 1) == '\0')
-		ft_printf(fd_out, "declare -x %s\"\"\n", env);
-	else
-	{
-		write(fd_out, "declare -x ", 11);
-		while (env[i] != '=')
-			write(fd_out, &env[i++], 1);
-		i++;
-		write(fd_out, "=\"", 2);
-		while (env[i])
-			write(fd_out, &env[i++], 1);
-		write(fd_out, "\"\n", 2);
-	}
-}
-
 void	do_export(t_pipex *data, char **env, char **cmd, int fd_out)
 {
 	int	i;
 
 	i = 0;
 	if (!cmd[1])
-	{
-		sort_strings(env);
-		while (env[i])
-			print_export(env[i++], fd_out);
-	}
+		just_export_cmd(env, fd_out);
 	else
 	{
 		i = 1;
@@ -180,6 +88,5 @@ void	do_export(t_pipex *data, char **env, char **cmd, int fd_out)
 			env = putstr_in_array(data, env, cmd[i++]);
 		}
 	}
-	data->env = env;
-	data->exitcode = 0;
+	set_exitcode_and_env(data, env);
 }
